Fix links and enhance documentation in README
Updated links to OpenJDK project and added documentation references.
Slim the README to the minimal required information and redirect for all other stuff to the https://sapmachine.io/docs as one single source of truth. That will ensure that we in future have only one single place to update information. Happy coding